split methods in utils into utils folder
