--- packages/components/src/templates/next/layouts/Collection/filterUtils/index.ts
@@ -1,2 +1,0 @@
-export * from "./constants"
-export * from "./getAvailableFilters"
