+++ packages/components/src/templates/next/layouts/Collection/utils/constants.ts
