+++ packages/components/src/templates/next/layouts/Collection/utils/getAvailableFilters.ts
