+++ packages/components/src/templates/next/layouts/Collection/utils/getCategoryFilter.ts
