+++ packages/components/src/templates/next/layouts/Collection/utils/getPaginatedItems.ts
@@ -1,0 +1,14 @@
+import type { ProcessedCollectionCardProps } from "~/interfaces"
+
+export const getPaginatedItems = (
+  items: ProcessedCollectionCardProps[],
+  itemsPerPage: number,
+  currPage: number,
+) => {
+  const normalizedCurrPage = Math.max(1, currPage)
+
+  return items.slice(
+    (normalizedCurrPage - 1) * itemsPerPage,
+    normalizedCurrPage * itemsPerPage,
+  )
+}
