+++ packages/components/src/templates/next/layouts/Collection/utils/getYearFilter.ts
