+++ packages/components/src/templates/next/layouts/Collection/utils/shouldShowCategory.ts
@@ -1,0 +1,7 @@
+import type { Filter } from "../../../types/Filter"
+import { FILTER_ID_CATEGORY } from "./constants"
+
+// use filters generated as the single source of truth
+export const shouldShowCategory = (filters: Filter[]): boolean => {
+  return filters.some((filter) => filter.id === FILTER_ID_CATEGORY)
+}
