+++ packages/components/src/templates/next/layouts/Collection/utils/shouldShowDate.ts
@@ -1,0 +1,7 @@
+import type { ProcessedCollectionCardProps } from "~/interfaces"
+
+export const shouldShowDate = (
+  items: ProcessedCollectionCardProps[],
+): boolean => {
+  return items.some((item) => item.lastUpdated)
+}
